[REF/IMP] introduce patchWithCleanup and refactor test helpers
damn, should have first done the refactoring, then the patchwithCleanup
thing
