+++ addons/web/static/tests/actions/client_action_tests.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
-import { click, legacyExtraNextTick } from "../helpers/utility";
-import { makeFakeRouterService } from "../helpers/mocks";
+import { click, legacyExtraNextTick } from "../helpers/utils";
+import { makeFakeRouterService } from "../helpers/mock_services";
 import { getLegacy } from "web.test_legacy";
 import { actionRegistry } from "../../src/actions/action_registry";
 import { viewRegistry } from "../../src/views/view_registry";
