+++ addons/web/static/tests/actions/close_action_tests.js
@@ -1,7 +1,6 @@
 /** @odoo-module **/
 
-import { nextTick } from "../helpers/index";
-import { legacyExtraNextTick } from "../helpers/utility";
+import { nextTick, legacyExtraNextTick } from "../helpers/utils";
 import { getLegacy } from "web.test_legacy";
 import { actionRegistry } from "../../src/actions/action_registry";
 import { viewRegistry } from "../../src/views/view_registry";
