+++ addons/web/static/tests/actions/effects_tests.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
-import { makeFakeUserService, nextTick } from "../helpers/index";
-import { click, legacyExtraNextTick } from "../helpers/utility";
+
+import { click, legacyExtraNextTick, nextTick } from "../helpers/utils";
 import { getLegacy } from "web.test_legacy";
 import { actionRegistry } from "../../src/actions/action_registry";
 import { viewRegistry } from "../../src/views/view_registry";
@@ -8,6 +8,7 @@ import { createWebClient, doAction, getActionManagerTestConfig } from "./helpers
 import { Registry } from "../../src/core/registry";
 import { NotificationContainer } from "../../src/notifications/notification_container";
 import { EffectContainer } from "../../src/effects/effect_container";
+import { makeFakeUserService } from "../helpers/mock_services";
 
 let testConfig;
 // legacy stuff
