+++ addons/web/static/tests/actions/url_action_tests.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
-import { makeTestEnv } from "../helpers/utility";
-import { makeFakeRouterService } from "../helpers/mocks";
+import { makeTestEnv } from "../helpers/mock_env";
+import { makeFakeRouterService } from "../helpers/mock_services";
 import { actionRegistry } from "../../src/actions/action_registry";
 import { viewRegistry } from "../../src/views/view_registry";
 import { doAction, getActionManagerTestConfig } from "./helpers";
