+++ addons/web/static/tests/helpers/mock_services.js
@@ -5,10 +5,9 @@ import { makePreProcessQuery, makePushState, routeToUrl } from "../../src/servic
 import { SIZES } from "../../src/services/ui_service";
 import { rpcService } from "../../src/services/rpc_service";
 import { localization } from "../../src/localization/localization_settings";
-import { patch, unpatch } from "../../src/utils/patch";
-import { registerCleanup } from "./cleanup";
 import { translatedTerms } from "../../src/localization/translation";
 import { computeAllowedCompanyIds, makeSetCompanies } from "../../src/services/user_service";
+import { patchWithCleanup } from "./utils";
 
 const { Component } = owl;
 
@@ -29,8 +28,7 @@ export const defaultLocalization = {
 };
 
 export function makeFakeLocalizationService(config) {
-  patch(localization, "localization.mock.patch", defaultLocalization);
-  registerCleanup(() => unpatch(localization, "localization.mock.patch"));
+  patchWithCleanup(localization, defaultLocalization);
 
   return {
     name: "localization",
@@ -143,52 +141,6 @@ export function makeFakeRPCService(mockRPC) {
     },
     specializeForComponent: rpcService.specializeForComponent,
   };
-}
-
-export function makeTestOdoo(config = {}) {
-  return Object.assign({}, odoo, {
-    browser: {},
-    debug: config.debug,
-    session_info: {
-      cache_hashes: {
-        load_menus: "161803",
-        translations: "314159",
-      },
-      currencies: {
-        1: { name: "USD", digits: [69, 2], position: "before", symbol: "$" },
-        2: { name: "EUR", digits: [69, 2], position: "after", symbol: "€" },
-      },
-      user_context: {
-        lang: "en",
-        uid: 7,
-        tz: "taht",
-      },
-      qweb: "owl",
-      uid: 7,
-      name: "Mitchell",
-      username: "The wise",
-      is_admin: true,
-      partner_id: 7,
-      // Commit: 3e847fc8f499c96b8f2d072ab19f35e105fd7749
-      // to see what user_companies is
-      user_companies: {
-        allowed_companies: { 1: { id: 1, name: "Hermit" } },
-        current_company: 1,
-      },
-      db: "test",
-      server_version: "1.0",
-      server_version_info: ["1.0"],
-    },
-    serviceRegistry: config.serviceRegistry,
-    mainComponentRegistry: config.mainComponentRegistry,
-    actionRegistry: config.actionRegistry,
-    systrayRegistry: config.systrayRegistry,
-    errorDialogRegistry: config.errorDialogRegistry,
-    userMenuRegistry: config.userMenuRegistry,
-    debugRegistry: config.debugRegistry,
-    viewRegistry: config.viewRegistry,
-    commandCategoryRegistry: config.commandCategoryRegistry,
-  });
 }
 
 export function makeMockXHR(response, sendCb, def) {
