+++ addons/web/static/tests/services/router_service_tests.js
@@ -5,10 +5,12 @@ import {
   parseHash,
   parseSearchQuery,
   redirect,
-  routeToUrl,
+  routeToUrl
 } from "../../src/services/router_service";
 import { patch, unpatch } from "../../src/utils/patch";
-import { makeTestEnv, nextTick } from "../helpers/index";
+import { makeTestEnv } from "../helpers/mock_env";
+import { nextTick } from "../helpers/utils";
+
 
 QUnit.module("Router");
