+++ addons/web/static/tests/services/rpc_service_tests.js
@@ -6,7 +6,9 @@ import { Registry } from "../../src/core/registry";
 import { notificationService } from "../../src/notifications/notification_service";
 import { rpcService } from "../../src/services/rpc_service";
 import { patch, unpatch } from "../../src/utils/patch";
-import { getFixture, makeDeferred, makeMockXHR, makeTestEnv, nextTick } from "../helpers/index";
+import { makeTestEnv } from "../helpers/mock_env";
+import { makeMockXHR } from "../helpers/mock_services";
+import { getFixture, makeDeferred, nextTick } from "../helpers/utils";
 
 const { Component, mount, tags } = owl;
 const { xml } = tags;
