+++ addons/web/static/tests/webclient/navbar_tests.js
@@ -1,16 +1,17 @@
 /** @odoo-module **/
 
-import { NavBar } from "../../src/webclient/navbar/navbar";
-import { click } from "../helpers/index";
-import { getFixture, makeTestEnv, nextTick } from "../helpers/utility";
-import { Registry } from "./../../src/core/registry";
 import { actionService } from "../../src/actions/action_service";
+import { browser } from "../../src/core/browser";
 import { hotkeyService } from "../../src/hotkey/hotkey_service";
-import { menuService } from "./../../src/services/menu_service";
 import { notificationService } from "../../src/notifications/notification_service";
 import { uiService } from "../../src/services/ui_service";
 import { patch, unpatch } from "../../src/utils/patch";
-import { browser } from "../../src/core/browser";
+import { NavBar } from "../../src/webclient/navbar/navbar";
+import { makeTestEnv } from "../helpers/mock_env";
+import { click, getFixture, nextTick } from "../helpers/utils";
+import { Registry } from "./../../src/core/registry";
+import { menuService } from "./../../src/services/menu_service";
+
 
 const { Component, mount, tags } = owl;
 const { xml } = tags;
