+++ addons/web/static/tests/webclient/webclient_tests.js
@@ -1,14 +1,15 @@
 /** @odoo-module **/
 
-import { WebClient } from "../../src/webclient/webclient";
-import { Registry } from "../../src/core/registry";
 import { actionService } from "../../src/actions/action_service";
-import { notificationService } from "../../src/notifications/notification_service";
-import { getFixture, makeTestEnv } from "../helpers/index";
-import { menuService } from "../../src/services/menu_service";
-import { fakeTitleService } from "../helpers/mocks";
+import { Registry } from "../../src/core/registry";
 import { hotkeyService } from "../../src/hotkey/hotkey_service";
+import { notificationService } from "../../src/notifications/notification_service";
+import { menuService } from "../../src/services/menu_service";
 import { uiService } from "../../src/services/ui_service";
+import { WebClient } from "../../src/webclient/webclient";
+import { makeTestEnv } from "../helpers/mock_env";
+import { fakeTitleService } from "../helpers/mock_services";
+import { getFixture } from "../helpers/utils";
 
 const { Component, tags, mount } = owl;
 const { xml } = tags;
